requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
anyio=3.6.2=pypi_0
apache-libcloud=3.7.0=pypi_0
appdirs=1.4.4=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.2.3=pypi_0
assertpy=1.1=pypi_0
asttokens=2.2.1=pypi_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.2=pypi_0
black=23.3.0=pypi_0
bleach=6.0.0=pypi_0
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.12.7=ha878542_0
certifi=2022.12.7=pypi_0
cffi=1.15.1=pypi_0
charset-normalizer=3.1.0=pypi_0
click=8.1.3=pypi_0
comm=0.1.3=pypi_0
contourpy=1.0.7=pypi_0
coolname=1.1.0=pypi_0
cudatoolkit=11.8.0=h37601d7_11
cupy-cuda113=10.6.0=pypi_0
cycler=0.11.0=pypi_0
debugpy=1.6.7=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
docker-pycreds=0.4.0=pypi_0
executing=1.2.0=pypi_0
fastargs=1.2.0=pypi_0
fastjsonschema=2.16.3=pypi_0
fastrlock=0.8.1=pypi_0
ffcv=1.0.2=pypi_0
fonttools=4.39.3=pypi_0
fqdn=1.5.1=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
gputil=1.4.0=pypi_0
idna=3.4=pypi_0
ipykernel=6.22.0=pypi_0
ipython=8.13.2=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.0.6=pypi_0
isoduration=20.11.0=pypi_0
isort=5.12.0=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
jsonpointer=2.3=pypi_0
jsonschema=4.17.3=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.2.0=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.3.0=pypi_0
jupyter-events=0.6.3=pypi_0
jupyter-server=2.5.0=pypi_0
jupyter-server-terminals=0.4.4=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=3.0.7=pypi_0
kiwisolver=1.4.4=pypi_0
ld_impl_linux-64=2.40=h41732ed_0
libffi=3.4.2=h7f98852_5
libgcc-ng=12.2.0=h65d4601_19
libgomp=12.2.0=h65d4601_19
libnsl=2.0.0=h7f98852_0
libsqlite=3.40.0=h753d276_1
libstdcxx-ng=12.2.0=h46fd767_19
libuuid=2.38.1=h0b41bf4_0
libzlib=1.2.13=h166bdaf_4
llvmlite=0.40.0=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=pypi_0
mistune=2.0.5=pypi_0
mosaicml=0.7.1=pypi_0
mypy-extensions=1.0.0=pypi_0
nbclassic=1.0.0=pypi_0
nbclient=0.7.4=pypi_0
nbconvert=7.3.1=pypi_0
nbformat=5.8.0=pypi_0
ncurses=6.3=h27087fc_1
nest-asyncio=1.5.6=pypi_0
notebook=6.5.4=pypi_0
notebook-shim=0.2.3=pypi_0
numba=0.57.0=pypi_0
numpy=1.24.3=pypi_0
opencv-python=4.7.0.72=pypi_0
openssl=3.1.0=hd590300_3
packaging=23.1=pypi_0
pandas=2.0.1=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pathspec=0.11.1=pypi_0
pathtools=0.1.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.5.0=pypi_0
pip=23.1.2=pyhd8ed1ab_0
platformdirs=3.5.0=pypi_0
prometheus-client=0.16.0=pypi_0
prompt-toolkit=3.0.38=pypi_0
protobuf=4.22.4=pypi_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
py-cpuinfo=9.0.0=pypi_0
pycparser=2.21=pypi_0
pydeprecate=0.3.2=pypi_0
pygments=2.15.1=pypi_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.19.3=pypi_0
python=3.10.10=he550d4f_0_cpython
python-dateutil=2.8.2=pypi_0
python-dotenv=1.0.0=pypi_0
python-json-logger=2.0.7=pypi_0
pytorch-pfn-extras=0.6.6=pypi_0
pytorch-ranger=0.1.1=pypi_0
pytz=2023.3=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=25.0.2=pypi_0
qtconsole=5.4.3=pypi_0
qtpy=2.3.1=pypi_0
readline=8.2=h8228510_1
requests=2.30.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
ruamel-yaml=0.17.23=pypi_0
ruamel-yaml-clib=0.2.7=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.12.2=pypi_0
send2trash=1.8.2=pypi_0
sentry-sdk=1.22.1=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=67.7.2=pyhd8ed1ab_0
sgdcollapse=0.0.0=dev_0
six=1.16.0=pypi_0
smmap=5.0.0=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.4.1=pypi_0
stack-data=0.6.2=pypi_0
terminado=0.17.1=pypi_0
terminaltables=3.1.10=pypi_0
threadpoolctl=3.1.0=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h27826a3_0
tomli=2.0.1=pypi_0
torch=1.11.0+cu113=pypi_0
torch-optimizer=0.1.0=pypi_0
torchaudio=0.11.0+cu113=pypi_0
torchmetrics=0.7.3=pypi_0
torchvision=0.12.0+cu113=pypi_0
tornado=6.3.1=pypi_0
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pypi_0
typing-extensions=4.5.0=pypi_0
tzdata=2023.3=pypi_0
uri-template=1.2.0=pypi_0
urllib3=1.26.15=pypi_0
wandb=0.15.2=pypi_0
wcwidth=0.2.6=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.5.1=pypi_0
wheel=0.40.0=pyhd8ed1ab_0
widgetsnbextension=4.0.7=pypi_0
xz=5.2.6=h166bdaf_0
yahp=0.1.0=pypi_0
